testes controller

+++ src/main/java/com/cinema/tickets/domain/service/impl/FilmeServiceImpl.java
@@ -1,11 +1,12 @@
 package com.cinema.tickets.domain.service.impl;
 
 import com.cinema.tickets.domain.collection.Filme;
+import com.cinema.tickets.domain.exception.BusinessException;
+
 import com.cinema.tickets.domain.exception.NotFoundException;
 import com.cinema.tickets.domain.repository.FilmeRepository;
 import com.cinema.tickets.domain.service.FilmeService;
 import com.cinema.tickets.domain.strategy.FilmeStrategy;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,15 +16,14 @@ import java.util.logging.Logger;
 @Service
 public class FilmeServiceImpl implements FilmeService {
 
-    Logger log= Logger.getLogger(FilmeServiceImpl.class.getName());
+    final Logger log= Logger.getLogger(FilmeServiceImpl.class.getName());
 
-    @Autowired
-    FilmeStrategy filmeStrategy;
+    private final FilmeStrategy filmeStrategy;
 
-    @Autowired
     private final FilmeRepository filmeRepository;
 
-    public FilmeServiceImpl(FilmeRepository filmeRepository) {
+    public FilmeServiceImpl(FilmeStrategy filmeStrategy, FilmeRepository filmeRepository) {
+        this.filmeStrategy = filmeStrategy;
         this.filmeRepository = filmeRepository;
     }
 
@@ -61,7 +61,7 @@ public class FilmeServiceImpl implements FilmeService {
         Optional<Filme> filmeOptional = this.filmeRepository.findById(filme.getId());
 
         if(filmeOptional.isEmpty()) {
-            throw new NotFoundException("Filme não encontrado");
+            throw new BusinessException("Filme não encontrado, ou dados inválidos");
         }
 
         log.info("Filme: " + filme.getTitulo() + " atualizado com sucesso");
@@ -71,7 +71,15 @@ public class FilmeServiceImpl implements FilmeService {
     @Override
     public void deleteById(String id) {
         log.info("Deletando filme pelo id: " + id);
+        Optional<Filme> filmeOptional = this.filmeRepository.findById(id);
+
+        if(filmeOptional.isEmpty()) {
+            throw new NotFoundException("Filme não encontrado");
+        }
+
+
         this.filmeRepository.deleteById(id);
+
         log.info("Filme deletado com sucesso");
     }
 }
